Due to issues with sharing the USART0_RX interrupt with the application, the bootloader now polls the UART to receive characters.
As the bootloader interrupt vector table is used instead of the application interrupt vector table, naked ISRs have been written for each interrupt which redirect to the application interrupt vector table. This could be done more effectively, but it would involve a custom interrupt vector table and writing the pre-main code. The additional latency introduced is four cycles. These ISR redirects are in "interrupts.h".

The bootloader now looks at the state of PD5. If this pin is high, the RS485 receiver is disabled, so the bootloader goes straight to the application.

At this stage, the bootloader could be considered complete. Tests have been performed and it performs as expected. There are inefficiencies that could be addressed, but they do not interfere with the behaviour of the code.